+++ src/main/java/ch/guessbb/sopraserver/repository/AchievementRepository.java
@@ -1,0 +1,8 @@
+package ch.guessbb.sopraserver.repository;
+
+import ch.guessbb.sopraserver.entity.Achievement;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AchievementRepository extends JpaRepository<Achievement, Long> {}