+++ src/main/java/ch/guessbb/sopraserver/repository/UserAchievementRepository.java
@@ -1,0 +1,8 @@
+package ch.guessbb.sopraserver.repository;
+
+import ch.guessbb.sopraserver.entity.UserAchievement;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserAchievementRepository extends JpaRepository<UserAchievement, Long> {}