+++ src/main/java/ch/guessbb/sopraserver/websocket/WebSocketConfig.java
@@ -1,7 +1,7 @@
 package ch.guessbb.sopraserver.websocket;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
@@ -13,12 +13,21 @@ import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerCo
 @EnableWebSocketMessageBroker
 public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
 
-    private final TopicSubscriptionInterceptor topicSubscriptionInterceptor;
+    private final WebSocketAuthInterceptor webSocketAuthInterceptor;
+
+    @Value("${rabbitmq.host}")
+    private String rabbitmqHost;
+
+    @Value("${rabbitmq.user}")
+    private String rabbitmqUser;
+
+    @Value("${rabbitmq.pass}")
+    private String rabbitmqPass;
 
 
     @Autowired
-    public WebSocketConfig(TopicSubscriptionInterceptor topicSubscriptionInterceptor) {
-        this.topicSubscriptionInterceptor = topicSubscriptionInterceptor;
+    public WebSocketConfig(WebSocketAuthInterceptor webSocketAuthInterceptor) {
+        this.webSocketAuthInterceptor = webSocketAuthInterceptor;
     }
 
 
@@ -33,13 +42,20 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
 
     @Override
     public void configureMessageBroker(MessageBrokerRegistry registry) {
-        registry.enableSimpleBroker("/topic");
+        registry.enableStompBrokerRelay("/topic")
+                .setRelayHost(rabbitmqHost)
+                .setRelayPort(61613)
+                .setClientLogin(rabbitmqUser)
+                .setClientPasscode(rabbitmqPass)
+                .setSystemLogin(rabbitmqUser)
+                .setSystemPasscode(rabbitmqPass);
+        //registry.enableSimpleBroker("/topic");
         registry.setApplicationDestinationPrefixes("/app");
     }
 
-    @Override
+    /*@Override
     public void configureClientInboundChannel(ChannelRegistration registration) {
-        registration.interceptors(topicSubscriptionInterceptor);
-    }
+        registration.interceptors(webSocketAuthInterceptor);
+    }*/
 
 }